RDCC-1952 refactored integration test cases
