--- src/integrationTest/java/uk/gov/hmcts/reform/cwrdapi/docs/SpringBootIntegrationTest.java
@@ -1,16 +1,0 @@
-package uk.gov.hmcts.reform.cwrdapi.config;
-
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.web.server.LocalServerPort;
-import org.springframework.test.context.junit4.SpringRunner;
-import uk.gov.hmcts.reform.cwrdapi.CaseWorkerRefApiApplication;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest(classes = CaseWorkerRefApiApplication.class, webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public abstract class SpringBootIntegrationTest {
-
-    @LocalServerPort
-    protected int port;
-
-}
