+++ src/integrationTest/resources/application.yml
@@ -1,5 +1,5 @@
 spring:
-  flyway.locations: classpath:db/migration
+  flyway.locations: classpath:db/testmigration
 
   jpa:
     show-sql: true
@@ -51,3 +51,4 @@ oidc:
   expiration: 14400000 # milliseconds 4 hours
 
 loggingComponentName: RD-Caseworker-Ref-Api
+userProfUrl: ${USER_PROFILE_URL:http://rd-user-profile-api-aat.service.core-compute-aat.internal}
